requirements.txt
pytest>=8.2.0
pytest-xdist>=3.6.1
pytest-rerunfailures>=14.0
allure-pytest>=2.13.5
selenium>=4.21.0
webdriver-manager>=4.0.2
requests>=2.32.3
python-dotenv>=1.0.1
pytest-check>=2.4.1